Feed logic with left middle and right section separated into a page container layout and the middle section will be used as a children prop in the app
